requirements.txt
# =====================================================================
# Satellite Garbage Dump Detection - Requirements
# Install via: pip install -r requirements.txt
# =====================================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Data Science & Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Web Application
streamlit>=1.28.0

# Geo-precise Map
folium>=0.14.0
streamlit-folium>=0.15.0

# GPS EXIF extraction
piexif>=1.1.3

# PDF Report generation
fpdf2>=2.7.4

# Real-time streaming (Pathway)
pathway>=0

# Multilingual Support
googletrans==4.0.0rc1
langdetect>=1.0.9

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
